refac: 재시도 로직 추가를 위해 resizeImage 로직은 utils로 분리
